requirements.txt
torch==1.12.1+cu113
torchvision==0.13.1+cu113
scikit-learn==1.0.2
pandas
numpy
Pillow==7.1.2
